README.md
This was a relatively small program when i first got it I was writing on this fro 3 days thought it was cool but soon was bored and back onto security I do believe if i can remember correctly I got some code from maybe the jarvis project? am not sure but i added to it got the microphone to work relatively good by setting the  setting the threshold and compensating for ambient noise in the room!
this application does quite a few things it was originally developed on archlinux so all commands may not work on other distros.

Maybe someday ill add either a easy way to edit the commands or just compile various commands and translate the commands o the right *nix system and have the application do fingerprinting on the system to pick the right syntax!!

add a database like sqlite for storage and retrieval of information!!!

if this is the right version i added it to remember your name and color and try to distinguish between two people with the same name

but then again i used no math really so the AI apart from google its shit!